requirements.txt
playwright>=1.40.0
openai>=1.3.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
lxml>=4.9.0
asyncio-throttle>=1.0.2 